EntityClassChooser now populated in separate thread
Window appears immediately, then tree is populated afterwards (consistent with
the UI principle of providing immediate feedback following a user operation).
Although for some reason the actual time taken to populate the treeview in the
separate thread is a fraction of what it was before, perhaps this is due to
sorting or some other issue apart from threads.