Removed RENDER_SCREEN render flag
This was used by a single shader, $OVERBRIGHT, which was rendered on a full-size
quad at the end of the rendering process in order to double the brightness of
the scene with a custom GL blend mode. This is now achieved with a simple
multiply in the fragment shader.

No analysis of whether this light scaling is required has been done; this is
just preserving the existing behaviour.