Fixed the incorrect renaming issue (#2733)
The key fix is at Namespace.cpp:261, where we need to check whether the imported
name conflicts with a name in the *destination* namespace, not the union of both
namespaces (which it always will, because we just built the union by adding all
of the new names).

Also includes further refactoring and cleanup.