Fix issue #3219: Preferences/Readable Editor failure
RE failed to save to fs_game_base if fs_game_base is not a direct child
of the D3-Folder. os::getContainingDir was used in the VFS to construct
the modpath, which does not work in this case, so it was replaced with
the new function os::getRelativePathMinusFilename, that retrieves the
folders down to the engine path.