Initial trivial test for Vector3
Boost.Test library is now a build dependency on Linux and is checked for in the
configure script.

The tests are known about by Automake and run automatically on Linux with the
"make check" command. There should be no effect on Windows at the moment as the
new files simply aren't referenced.