Fixed inappropriate vertex stride when rendering wireframe patches
Introduced new render::VertexBuffer helper class, which accepts vertices in
batches (possibly using a stride value) and stores them internally for later
submission to OpenGL in a more efficient way. In future this class can be
extended to use VBOs, provide support for colour and texture coordinate
submission and be more generally useful throughout the rendering code.

Hopefully this will fix the rendering issues on nVidia cards but this needs to
be tested on the respective hardware.