VertexBuffer uses VBO for vertex submission
Unlikely to be any performance improvement at present because the only user of
VertexBuffer (RenderablePatchWireframe) is creating and destroying the buffer
with each render call. However the code is more compatible with future OpenGL
versions where the only acceptable way to submit geometry is through VBOs.

Note use of offsetof() macro in ArbitraryMeshVertex; may require fix-up on
MSVC.