Avoid unnecessary VBO allocations when re-rendering patches
Introduced VertexBuffer::replaceVertexData() which uses glBufferSubData to
replace the VBO's data store with a vector of vertices from another
VertexBuffer of the same size or smaller. A new allocation is only required
when the number of vertices in the patch mesh changes (e.g. after a significant
change in geometry).