Fix more plane representation madness.
The formulae to derive the frustum planes from the viewprojection matrix
are correct but refer to a,b,c,d plane format. DarkRadiant stores a,b,c
and dist... as long as all the code assumes that the value in Plane3::_dist
is actually d everything is fine, but since my projected light changes
this is no longer the case.