Fix TAB event not being processed (sometimes).
In Windows, TAB events only result in a keydown event if the event
object in question has the wxWANTS_CHARS flag set. Otherwise (which is
the default for wxPanels) the keypress results in wxNavigationEvents
which are not handled by the GlobalKeyEventFilter. The specific fix for
this was to enable the flag for the status bar panels which were in
focus in the reported case.