Force rendering of selected nodes that would otherwise be hidden
With grouping and the filter system situations can occur where parts of
a group are filtered out but are selected through group peers. Instead
of preventing the selection, of the whole group or selecting just the
visible parts, we force the hidden group parts to be visible such that
the user can see what's going to be affected by the next operation.