Resolve #4551 in wxGTK after fixing it in wxMSW
In Window Layouts with the GroupDialog in its own Window the attempt to
Reparent() the page widget to the wxNotebook fails and the page gets
cramped into the upper left corner. I can't figure the heck out why this
happens, but it works when inserting a wxPanel and reparenting the page
to that instead.