Add initial snapcraft build file
Snaps are a relatively new (but well-supported) packaging format for
distributing applications on Linux, which should allow a single package to be
usable on a wide variety of distributions without running into the usual
problems with shared library dependencies.

This initial file allows a basic .snap package to be built, however the
contained binary does not start (it seems to be unable to find its bundled
libraries).