Add a configure option to enable relocation
Self-contained packaging systems like Snappy require binaries to be
relocatable, in the sense of not containing any hard-coded pathnames like
/usr/share/darkradiant. This allows the package manager to install the binary
at an arbitrary location which may not be known at the time the software was
compiled.

The configure script now has a new option '--enable-relocation' which sets an
ENABLE_RELOCATION #define when building, although as yet there are no code
changes and this define has no effect.