Remove ArchiveLoader module altogether
All the Pk4ArchiveLoader ever did was construct an instance of ZipArchive which
was used in a single place (Doom3FileSystem). Now that both modules are merged
into a single directory of the main binary, the Doom3FileSystem can simply
construct a ZipArchive itself without needing a separate module just for this.

The iarchive.h header remains because it defines certain data structures that
are used elsewhere, however there is nothing module-related in this header any
more.