Remove persistence of EmbeddedLayout pane positions
A bug in the persistence code is causing a progressive shift of the pane
positions with each successive session, moving the splitters to the right and
downwards until the 2D view and group widgets are hardly visible. With this
code disabled, the splitters are left at their default position, which (on
Linux at least) seems to be 50% of the widget size. This is not optimal but at
least is consistent between sessions.