UNIX manpage is now generated from an AsciiDoc source file
AsciiDoc is a text formatting language similar to MarkDown, which can be used
to generate a variety of other documentation formats, including manpages. This
makes the manpage much easier to edit since there is no need to use the
esoteric ROFF syntax directly. A Makefile contains the necessary commands to
generate the manpage from the AsciiDoc source.