Use the 'header' table style where appropriate
The 'header' style gives bold text and a darker background, which makes header
cells more obvious, and avoids the need to manually format each piece of header
text with inline bold markup.

Since a column can only have one style, this replaces the 'asciidoc' style
which was used to allow double-colon image declarations. This is not a problem,
as single-colon images work fine without the 'asciidoc' style, and the column
specifier can be used to set the desired center alignment.