Add an extremely trivial unit test for Doom3FileSystem
All the test does is construct an instance of Doom3FileSystem and confirm its
module name and dependencies. This required moving the StaticModule instance
owner into a separate .cpp file so it could be excluded from the test binary
(otherwise this would pull in a whole load of cascading dependencies starting
with the modulesystem).