Fix broken filter rule editor on Linux
Due to some combination of quirks with the GTK-based implementation of
wxWidgets, the custom filter rule editor was not working at all (edits were
visible in the GUI but not propagated to the filter object itself, and were
therefore reverted the next time the dialog was shown).

The FilterEditor is now simplified and using the standard wxDataViewListCtrl
widget which provides a basic list of items that can be displayed and edited;
previously it was using the custom TreeModel class which is overkill for this
situation (and possibly many other use cases in DarkRadiant).

Also removed a couple of other TreeModels and related classes which seemed to
be entirely unused.