Hide entire camera toolbar in response to preference
The preference to hide the camera toolbar was only hiding the first four icons
(related to render modes) while leaving other icons visible. This was because
the visible toolbar is actually composed of two wxWidgets toolbars: the
CamToolbar and the MiscToolbar. The new slot controls the visibility of both of
these items.