Remove close button from modal dialogs
A user reported that accidentally closing dialogs such as the Objectives editor
with the top-right close button can result in data loss, because the dialog
treats this as "Cancel" and does not prompt to save changes. Now the close
button is removed, so a user must explicitly choose to click one of the main
buttons at the bottom.

I don't know for sure whether the wxSYSTEM_MENU style is needed, but since it
was part of the wxDEFAULT_DIALOG_STYLE and I'm not in a position to test its
removal (it only has an effect on Windows according to the docs), it is left in
for now.