Minor tweak to DifficultyEditor .fbp file
Set the global properties in the .fbp file to specify that we are generating
XML code (not C++ which is the default), and that the output file should be
named 'difficultyeditor'. This causes wxFormBuilder to automatically write the
correct .xrc file when pressing F8.

Also removed the internal border inside the left-hand tree panel, so that the
inserted tree view is horizontally aligned with the buttons.