Minor refactor of DifficultyEditor construction
Merge the two separate loops into a single loop which constructs each
DifficultyEditor, adds it to the internal list and then inserts it as a
choicebook page. This means that DifficultyEditor no longer needs to store and
return its own label since the text is already known in the population loop.