Working upload of uncompressed RGB DDS images
If the DDS header does not have a FOURCC indicating compressed data, we make a
lookup into another map which associates DDS bit depths (24 for RGB, 32 for
RGBA) with OpenGL memory layouts (GL_RGB or GL_RGBA), and pass an additional
parameter to setFormat() indicating that the image is uncompressed and should
be uploaded with glTexImage2D rather than glCompressedTexImage2D.

The vertexcolors_break_shading test map now renders correctly including the
uncompressed DDS placeholder texture, which is in RGB8 (24 bit) format. No
testing of RGBA8 textures so far.