Config directory on Linux now complies with FreeDesktop standard
Instead of using $HOME/.darkradiant as the settings directory, we are now using
$HOME/.config/darkradiant in accordance with modern standards. Currently this
also includes the log file, which is permissible but not ideal since this is
not configuration data ($HOME/.cache/darkradiant would be a better location for
temporary output like log files).

Also fix a possible (although unlikely) crash bug: assigning the result of
getenv() directly to a std::string is dangerous because getenv() can return
NULL and it is (surprisingly) undefined behaviour to construct a std::string
from a NULL pointer. This would only occur in released builds if either $HOME
(Linux) or %APPDATA% (Windows) are missing, which seems exceptionally unlikely.