RenderableCollector only has a single addRenderable method
RenderableCollector has finally discovered the cutting-edge C++03 features of
"null pointers" and "default arguments", and now exposes only a single
addRenderable method with optional pointers to a LightList and an IRenderEntity
object.

Previously there were three separate methods with different combinations of
IRenderEntity and LightList parameters, resulting in pointless duplication of
code in the implementing subclasses.