RenderableCollector::addRenderable() takes a Shader reference
Since RenderableCollectors are ephemeral objects which only last for a single
render operation, there is no need for them to receive a full shared pointer to
a shader and possibly incur the expense of incrementing the reference count.

In practice this probably makes no difference to performance since every
addRenderable() implementation was just dereferencing the shared_ptr
immediately and never storing it, but since addRenderable() is called a great
many times for each frame, we might as well make the interface as streamlined
as possible.