Add a proper interface for IGameConnection
Previously the IGameConnection interface was empty, but the code which used the
module actually did a dynamic_cast down to the GameConnection implementing
class and used its methods directly, thereby breaking the module abstraction.

Now, the methods used by the UI code have been added as pure virtuals in the
IGameConnection interface, eliminating the need for the dynamic downcast.