Adjust GL raster text and render statistics line
Increase the font to 14 pixels (from 12) and format the render statistics on a
single line. Since we're still using evil hard-coded pixel sizes there is a
limit to how much we can tweak this (the same font is used for drawing other
text such as grid coordinates and brush bounds).