Render statistics track front-end and back-end times separately
Store separate time values for render front-end and back-end stages, and
display these separately on the stats line along with the total and the
calculated fps. Currently on my Linux system this shows that the majority of
time is spent in the back-end, with the initial scenegraph traversal happening
relatively quickly.