Initial code to support lighting calculations by RenderableCollector
Added a new method to RenderableCollector to receive renderables under the
control of a LitObject, which will be used to calculate light intersections in
a subsequent pass rather than requiring the object to maintain its own list of
intersecting lights. Added initial code in CamRenderer to implement this method
and calculate light intersections, although nothing is using it yet.