Initial debugging code for light projection matrix
If DEBUG_LIGHT_MATRIX is set at compile time, the effect of the _localToTexture
matrix will be shown as applied to certain key points: light origin, target,
and the top-left and bottom-left corners of the target plane. This is more
useful than examining the matrix itself which is difficult to interpret for
anything other than trivial matrices.