StaticModel is now using addLitRenderable() to submit geometry
Also updated the RenderPreview to use CamRenderer as its RenderableCollector;
this is necessary because CamRenderer now contains the essential logic for
calculating light intersections, without which the RenderPreview will only show
a black image.

Because CamRenderer has been refactored into a separate header in libs/render,
some of the render statistics are no longer working (because CamRenderer no
longer has access to code in the radiant module). This will need to be fixed
later.