Main binary is now buildable with CMake
Enable the '-z defs' linker which detects undefined symbols in shared objects,
and used this information to add the required link options to our shared
libraries. Because these link options are declared PUBLIC in CMake, they
automatically propagate through to whatever links against the built library, so
we don't need to manually repeat all of the link options when building the main
binary.

Also added some missing source files, including CPP files in subdirectories of
wxutil, and the PicoModel C files in radiantcore.

C++ standard version is increased to 17 to allow for std::filesystem; if this
proves problematic the fallback to boost::filesystem can be brought across from
the Automake build system.