Install binary and its shared library dependencies
Install the darkradiant binary and the shared libraries it links against
(wxutil, math etc), and set a relative -rpath so that the executable can find
its libraries. Unlike with Automake, the use of "$ORIGIN/.." just works without
needing to figure out how to escape the syntax through two levels of shell
scripts plus the Makefile.

The binary does now start to run without any linker error, but the application
does not yet work because it cannot load the main module (probably because it
is called 'libradiantcore.so' not 'radiantcore.so').