Pass package library path to dh_shlibdeps
Using the -l option to dh_shlibdeps we can add the /usr/lib/darkradiant
directory to the search path for shared libraries, reducing the number of
warnings like "Cannot find libwxutil.so...". There are still warnings like
"Cannot extract name and version info from libwxutil.so" which seem to be
unavoidable.