CMake build now has support for non-relocatable builds
Although the previous build was not defining ENABLE_RELOCATION, it was treating
the build as relocatable by default, setting an RPATH based on $ORIGIN and
never defining PKGLIBDIR, PKGDATADIR etc in the source code. Although this
works fine on Debian, the assumption that the libraries are ALWAYS available at
$ORIGIN/../lib/darkradiant might not be suitable for all systems.

We now expose an additional CMake option ENABLE_RELOCATION which defaults to
ON; by setting this to OFF the build becomes the more usual Unix-style build
with hard-coded absolute data and library paths. In addition, when making a
relocatable build the expected ENABLE_RELOCATION macro is also defined in
config.h.