Install drtest binary
drtest binary is installed to the 'bin' directory alongside the main
darkradiant binary (this is different from the Automake build, which installed
drtest to a subdirectory of $pkglibdir).

This in turn exposed a link issue with libradiantcore.so, which was using the
same '$ORIGIN/../lib/darkradiant' rpath as the main binary, which is actually
incorrect for libradiantcore which is in the lib/darkradiant/modules directory.
This did not matter when libradiantcore was being loaded from darkradiant, but
caused a runtime link failure with drtest.