Fix assertion on startup in Regular layout
Assertion caused by a recursive wxYield call: first in
LongRunningOperationHandler::onMessage and then in
RegularLayout::restoreStateFromPath.

The wxWidgets warn against use of wxYield presumably because it can cause
unexpected problems like this, so for now I have just removed the wxYield call
in RegularLayout without replacing it with anything. The startup behaviour
looks OK on Linux but if there are problems caused by window size, perhaps the
use of another function like ProcessPendingEvents() would be safer.