Fix unusable spin controls in Vector3PropertyEditor
Default hard-coded minimum size of 75 pixels was too small for the current
implementation of wxSpinCtrl on GTK (at least with my desktop theme) and
rendered the controls unusable. We now set a minimum size based on the range of
the numeric value, which might be a little too large but at least it's not a
hard-coded guess.