Avoid hard-coded '../lib' in ApplicationContextBase
Another hard-coded library directory lurking in ApplicationContextBase was
continuing to break relocatable builds on systems which default to 'lib64' as
their library directory name. Although the RPATH was fixed in bf5eaaa9264d1763,
the relative path is also needed when we dynamically load the modules.

The CMake script now writes a RELATIVE_LIBDIR variable into config.h which
contains the relative part of the library directory (i.e. starting with ".."),
and this RELATIVE_LIBDIR is now used both to set the RPATH in CMakeLists.txt
and in ApplicationContextBase to load the modules.