Avoid activating camera sync toggle if connection failed
Introduce a new method IEventManager::addAdvancedToggle() which adds a toggle
callback that returns a bool to indicate whether the toggle succeeded or not.
If it returns false, the state is not changed and the menu option does not
appear checked. This new advanced toggle is used by the 'Game position follows
camera' menu item, which no longer appears checked after failing to connect to
the game process.