CamWnd no longer implements ICameraView
CamWnd was only ever a "forwarding wrapper" for ICameraView, with public
methods which did nothing other than pass the call directly to the equivalent
method on the contained ICameraView::Ptr member.

This is now simplified by removing the ICameraView interface from CamWnd
entirely, and instead passing the contained ICameraView object to the
constructor of CameraMouseToolEvent (which also now needs a separate
IInteractiveView reference for the MouseToolEvent parent class to expose).

The ICameraView forwarding methods are mostly removed from CamWnd now, except
for a few which are still used explicitly (by CamWnd and other classes).

Also renamed the IMainFrame's toolbar enum values from "HORIZONTAL" and
"VERTICAL" to "TOP" and "LEFT" to make clear that they refer to specific
application toolbars, not possible toolbar orientations.