Add a toolbar button for "Sync camera back" function
This button is next to the existing camera sync toggle button, and corresponds
to the "Move camera to current game position" menu item. Both buttons' icons
are tweaked to be somewhat similar to one another (a blue eye above an arrow)
to make it clear that their functionality is closely related.