Attach entities at correct position
Attached light entities are now correctly parented to the main entity, and
their attachment offset is encoded into the attached entity's localToParent()
matrix. Some changes were required to Light which has previously always assumed
that its "origin" key IS the light position - we now transform the local origin
by localToWorld() which takes into account both the "origin" key and any
transformation applied to the light entity or its parent.

Attached lights are now appearing in the correct position and tests pass, but
the attached light does not move when the parent entity is dragged.