#5534: EntityClassChooser is no longer a static singleton
Since populating the tree for this dialog is very quick, there is no benefit in
keeping a singleton around. Creating a new dialog each time requires much less
code and also fixes the major issue described in #5534, namely that the dialog
is always empty on its second showing.

There is still a problem with the search feature (at least on Linux), which
results in the tree view becoming empty as soon as any text is entered in the
search box. However it is at least now possible to resume normal service by
closing the dialog and showing another one.