Add test for light wireframe shader
Since the Shader interface doesn't expose very much information that can be
examined in tests, the shader system now attaches an internal dummy Material to
any shader which doesn't have a Material of its own. Currently this internal
Material only exposes the name used to construct the shader, and it is this
which is tested to confirm that the light wireframe shader matches the
editor_color value.

In order to make it easier to construct this new InternalMaterial class, all
public virtuals on the Material interface are now non-pure, with empty default
implementations. This in turn required a couple of method signatures to be
changed to return by value rather than by const reference.