Add an experimental AUI-based layout
This layout approximates the Embedded layout but uses the wxWidgets AUI system
to manage the components, allowing for panels to be moved, docked and floated
as required by the user. Currently the basic layout works and the panels are in
approximately the right positions, but the panels are very thin and have to be
dragged out to fill the whole window, and there is no way to get them back if
the close button is clicked.