AUI panels now take up the entire window
Use a nasty hack to workaround AUI's inability to take into account BestSize()
when initially adding a docked pane. The panes are all added with a minimum
size of approximately half the window, then those minimum sizes are reset to a
much smaller value after the first Update(). This results in panels which take
up the entire window to begin with, but can still be shrunk by the user if
required.