2D view is now the CenterPane in AUI layout
The AUI Center pane is a non-floatable pane which takes up all the space in the
window not occupied by docked panes. Previously the 2D view was docked in the
Right position, which tended to leave a large empty area in the middle of the
window which would only be filled if the left and right panes were manually
resized.

The disadvantage of using the Center position is there is no possibility to
float the 2D view, but since we can create a new standalone window from the
View menu, this seems an acceptable compromise to avoid an annoying empty area.