Remove broken EntityClassAttribute::inherited flag
Ever since the fix for #5546 eliminated copying of inherited attributes from
parent entity classes into their children, the 'inherited' bool was rendered
useless, which in turn broke certain functionality such as the colour-coding of
inherited spawnargs in the EClassTree.

The inherited flag is now removed entirely from EntityClassAttribute, and
replaced by an extra bool passed to the AttributeVisitor when visiting entity
class attributes.