Add missing image dimension reads to Quake 3 legacy brush parser
I believe there is no exporter for that brush format so that would mean the
Q3 legacy brushes would be converted to Q3 non-legacy brushes on write,
if there is no bug on the writing part. That other brush format is compatible
with NetRadiant and the q3map2 map compiler so that would not be a big issue.

See that thread for work-in-progress status with the Q3-map based Unvanquished game:

- https://forums.unvanquished.net/viewtopic.php?f=9&t=2121