Remove string sharing mechanism from EntityClassAttribute
Sharing of strings was added to solve a specific problem with memory usage when
attributes (including those with long strings like editor_usage) were copied
into hundreds of child entities. Since we are no longer copying inherited
attributes but allowing children to access them directly, this mechanism is
largely redundant.

There might be a *slight* increase in memory usage (compared to the previous
commit) now that the EntityAttributeMap is storing name strings twice (once in
the key, and once in the EntityClassAttribute::_name member), but I assume this
is OK since the overall number of EntityClassAttributes is drastically reduced
due to the lack of copying, and the lengthy editor_usage values are not
duplicated.