Add a string::ILess functor for case-insensitive containers
Replace about 8 separate implementations of a std::less functor using
string_compare_nocase() with a single functor defined in string.h itself.

This is covered by tests in a new BasicTest suite which can be used for testing
simple core functions which do not require the construction of a complete
Radiant environment.