Remove light brightness doubling in shader
For some reason there was a hard-coded scale factor of 2.0 applied to all
lights in the shader (this is over and on top of the 2x scaling via the
lightScale setting in the .game file, so 4x brightness in total), resulting in
lights appearing far too bright and in many areas completely washed out.

The brightness now appears much closer to that seen in game (possibly slightly
darker because the game is providing some brightness and gamma scaling of its
own, which is not replicated in DarkRadiant).