Colour4 stores floats rather than doubles
Using doubles in vectors may be necessary for accuracy when storing brush and
patch geometry, but using doubles just to store OpenGL colours is a waste of
bytes.

Also add some tests for Vector3 and Vector3f packing, ensuring that the values
are contiguous and can be passed to functions like glUniform3fv().