Merge getRotationAboutZ degree and radian versions
Instead of two separate methods, there is now a single templated method
getRotationAboutZ which accepts a class which must define toRadians(). Two such
classes are added: math::Degrees and math::Radians, both of which have explict
constructors (so calling code must be clear about whether the provided angle is
in degrees or radians).