Test difference between affine inverse and full inverse
Both inverses should be the same when the matrix contains only an affine
transformation, and different if the matrix contains a projective
transformation.

Also add inline operator* and operator- for Matrix4, which surprisingly did not
already exist (multiplications are all implemented using the explicit
multiplyBy()/getMultipliedBy() methods, and I guess matrix subtractions aren't
a common requirement for computer graphics).