Move clipping functions out of Matrix4
clipPoint/clipTriangle/clipLine are utility functions that use a Matrix4 to
transform points, but do not access any private members of Matrix4 and
therefore do not need to be methods. These are now free functions in the
BestPoint.h header, near to the code which is using them (here and in
SelectionVolume.h).