Remove unsafe Matrix4 pointer in TransformedRenderable
TransformedRenderable was storing a non-owning C pointer to each renderable
object's transform matrix, which is inherently unsafe if any object has
returned a matrix calculated on the fly (i.e. an rvalue) rather than by
reference. Storing a copy of the matrix might be very slightly slower but
avoids this potential source of bugs.