Fix a disappearing brush issue
Commit e71c0c156d9bd1c5862b introduced a rendering bug causing brushes to
disappear in some circumstances. The changes to renderAllContained() were
excessive - there is nothing wrong with using a Matrix4 pointer locally in this
method, since the pointed Matrix4 object is stored in each
TransformedRenderable which outlives the local pointer - and the modified code
failed to correctly process the first renderable (when transform == nullptr).