Fix tests in out-of-source CMake build
We can't assume that the built core module can be found in the
./test/../radiantcore directory, because with an out-of-source build the shared
libraries do not appear in the source directory at all. With a separate build
directory we need the old behaviour of looking in PKGLIBDIR/modules, however
this doesn't work with in-source builds.

Fortunately getLibraryPaths() returns a list of directories, so we can just add
both possibilities to the list and get functional tests with both in-source and
out-of-source builds.