Remove hard-coded text colouring in Entity Inspector
The hard-coded colours assumed a light background, but the default black text
is very difficult to see with a dark theme applied. Instead of forcing colours
to distinguish entity properties from inherited properties, we now render the
inherited properties in italic text while leaving colours to be chosen by the
theme.