Upload uncompressed DDS in BGR rather than RGB order
It is unclear why DDS data is appearing in BGR order, since nothing in the
header or the specification suggests that data should be stored in this format.
However the game code also does the same swizzling, so this seems to be the
expected way to handle uncompressed DDS.

This probably will not affect most users since the use of uncompressed DDS is
rare.