Workaround for failure to render NPOT DDS mipmaps
Check the GL errors, and if there is an error after uploading a mipmap for
detail level > 0, fall back to calling glGenerateMipmap() after warning on the
console. This resolves the failure to render the NPOT texture, but it does not
fix the underlying issue with correctly uploading the file's mipmaps.